+++ app/forms/transfer/task/conditions_met_task_form.rb
@@ -1,0 +1,3 @@
+class Transfer::Task::ConditionsMetTaskForm < BaseTaskForm
+  attribute :confirm_all_conditions_met, :boolean
+end