+++ db/migrate/20230807105235_add_all_conditions_met_task_to_transfers.rb
@@ -1,0 +1,5 @@
+class AddAllConditionsMetTaskToTransfers < ActiveRecord::Migration[7.0]
+  def change
+    add_column :transfer_tasks_data, :conditions_met_confirm_all_conditions_met, :boolean
+  end
+end